Move camera files to camera folder
