+++ trview.app/Camera/Camera.cpp
