+++ trview.app/Camera/Camera.h
