+++ trview.app/Camera/CameraControls.cpp
