+++ trview.app/Camera/CameraControls.h
