+++ trview.app/Camera/CameraMode.cpp
