+++ trview.app/Camera/CameraPosition.cpp
