+++ trview.app/Camera/CameraPosition.h
