+++ trview.app/Camera/FreeCamera.cpp
