+++ trview.app/Camera/FreeCamera.h
