+++ trview.app/Camera/ICamera.cpp
