+++ trview.app/Camera/ICamera.h
