+++ trview.app/Camera/OrbitCamera.cpp
