+++ trview.app/Camera/OrbitCamera.h
