+++ trview.app/ViewerUI.h
@@ -16,8 +16,8 @@
 #include "SettingsWindow.h"
 #include "Flipmaps.h"
 #include "Neighbours.h"
-#include "CameraControls.h"
-#include "CameraPosition.h"
+#include "Camera/CameraControls.h"
+#include "Camera/CameraPosition.h"
 #include "PickInfo.h"
 #include "PickResult.h"
 #include "ContextMenu.h"
