+++ trview/Entity.cpp
@@ -6,7 +6,7 @@
 
 #include <trview.app/ILevelTextureStorage.h>
 #include "IMeshStorage.h"
-#include <trview.app/ICamera.h>
+#include <trview.app/Camera/ICamera.h>
 #include <trview.app/Mesh.h>
 #include <trview.app/TransparencyBuffer.h>
